fixed stream
