+++ FrenchGenderTrainer/Data/AspiratedH.swift
@@ -1,0 +1,11 @@
+//
+//  AspiratedH.swift
+//  FrenchGenderTrainer
+//
+//  Created by KATHRIN AYER on 1/21/20.
+//  Copyright © 2020 happierplaces. All rights reserved.
+//
+
+import Foundation
+
+
