+++ FrenchGenderTrainer/GameView.swift
@@ -12,6 +12,18 @@ import RxCocoa
 import AVKit
 
 class GameView: UIViewController {
+  private enum SpeechEnum: String {
+    case male
+    case female
+
+    func voiceProfile() -> String {
+      switch self {
+      case .male: return "fr"
+      case .female: return "fr-ca"
+      }
+    }
+  }
+
   @IBOutlet var leftButton: UIButton!
   @IBOutlet var rightButton: UIButton!
   @IBOutlet var wordLabel: UILabel!
@@ -23,6 +35,8 @@ class GameView: UIViewController {
   private var frenchWord: FrenchWord?
   private var labelObservable = BehaviorRelay<String>(value: "Start")
 
+  private let speechSynthesizer: AVSpeechSynthesizer! = AVSpeechSynthesizer()
+
   override func viewDidLoad() {
     super.viewDidLoad()
 
@@ -30,38 +44,59 @@ class GameView: UIViewController {
   }
 }
 
+// MARK: - Speech
+extension GameView {
+  private func speak(word: String) {
+    let speechUtterance = AVSpeechUtterance(string: word)
+    speechUtterance.voice = AVSpeechSynthesisVoice(language: "fr")
+
+    speechSynthesizer.speak(speechUtterance)
+  }
+}
+
+// MARK: - Verify Step
+extension GameView {
+  private func createWordPhrase(word: FrenchWord) -> String {
+    let string = word.french
+
+    return string ?? ""
+  }
+}
+
 // MARK: - Data Stream
 extension GameView {
-  private func newWord() {
+  private func newWord() -> FrenchWord? {
     // TODO: dataFacade.getRecords should be able to throw
     // TODO: if dataFacade fails, retry
     self.frenchWord = dataFacade.getRecords(fetchLimit: 1, predicate: nil)[0]
 
     guard let frenchWord = self.frenchWord else {
       // TODO: Create meaningful error
-      return
+      return nil
     }
     self.labelObservable.accept(frenchWord.french!)   //  = frenchWord.french!
 
     print(frenchWord)
+    return frenchWord
   }
 
   private func initObservation() {
     let gameViewModel = GameViewModel.init(leftButton: leftButton, rightButton: rightButton)
 
-    let interactionObservable = Observable
-      .combineLatest(gameViewModel.stateStream,
-                     gameViewModel.buttonStream.distinctUntilChanged())
+//    let interactionObservable = Observable
+//      .combineLatest(gameViewModel.stateStream,
+//      gameViewModel.buttonStream)
 
-    let interactionStream = interactionObservable
-      .subscribe(onNext: { (state, button) in
+    let interactionStream = gameViewModel.stateStream
+      .subscribe(onNext: { (state, button16) in
+//
         switch state {
         case .present:
-          self.newWord()
+          // TODO: if newWord returns nil, try again
+          self.frenchWord = self.newWord()
         case .explain:
           print("is it correct? show correct, else show explanation")
         case .verify:
-          let button16 = Int16(button)
           if button16 == self.frenchWord?.gender {
             print("CORRECT")
           } else {
@@ -73,6 +108,7 @@ extension GameView {
       }, onCompleted: {
         print("complete")
       })
+
     let labelSubscription = labelObservable
       .subscribe(onNext: { (string) in
         self.wordLabel.attributedText = NSAttributedString.init(string: string)
